backend/certificados-service/requirements.txt
# Certificados Service - Geração de Certificados para Eventos
Django==5.0.7
djangorestframework==3.15.2
django-cors-headers==4.3.1
python-decouple==3.8
psycopg2-binary==2.9.7
celery==5.3.4
redis==5.0.1
WeasyPrint==60.2
Pillow==10.0.1
requests==2.31.0
python-dateutil==2.8.2
reportlab==4.0.4
jinja2==3.1.2